fork_backend/pyproject.toml
[project]
name = "fork_backend"
version = "0.1.0"
description = "A calorie tracking app"
readme = "README.md"
requires-python = ">=3.12.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.14.1",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.128.0",
    "openfoodfacts>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pgvector>=0.4.2",
    "psycopg2>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pylint>=4.0.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
